tests: make treesitter its own category
